+++ packages/choiceguide-results/vite.config.ts
@@ -1,15 +1,18 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 import path from 'path';
+import { prefix } from '../lib/prefix';
 
 export default defineConfig(({ command }) => {
     if (command === 'serve') {
         return {
             plugins: [react()],
+            css: prefix()
         };
     } else {
         return {
             plugins: [react({ jsxRuntime: 'classic' })],
+            css: prefix(),
             build: {
                 lib: {
                     formats: ['iife'],
