+++ packages/resource-overview/vite.config.ts
@@ -1,8 +1,6 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
-
-import prefixer from 'postcss-prefix-selector';
-import autoprefixer from 'autoprefixer';
+import {prefix} from '../lib/prefix';
 
 
 // https://vitejs.dev/config/
@@ -11,41 +9,13 @@ export default defineConfig(({ command }) => {
   if (command === 'serve') {
     return {
       plugins: [react()],
+      css: prefix()
     };
     // During build, use the classic runtime and build as an IIFE so we can deliver it to the browser
   } else {
     return {
       plugins: [react({ jsxRuntime: 'classic' })],
-      css: {
-        postcss: {
-          plugins: [
-            prefixer({
-              prefix: '.openstad',
-              transform(prefix, selector, prefixedSelector, filePath, rule) {
-                if (selector.match(/^(html|body)/)) {
-                  return selector.replace(/^([^\s]*)/, `$1 ${prefix}`);
-                }
-
-                if (selector.match(/^(:root)/)) {
-                  return selector.replace(/^([^\s]*)/, `${prefix}`);
-                }
-    
-                if (filePath.match(/node_modules/)) {
-                  return selector;
-                }
-    
-                const annotation = rule.prev();
-                if (annotation?.type === 'comment' && annotation.text.trim() === 'no-prefix') {
-                  return selector;
-                }
-    
-                return prefixedSelector;
-              },
-            }),
-            autoprefixer({})
-          ],
-        }
-      },
+      css: prefix(),
       build: {
         lib: {
           formats: ['iife'],
